backend-flask-pytorch/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.16.1
Flask==2.3.2
Flask-Cors==5.0.0
fonttools==4.54.1
fsspec==2024.9.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
Pillow==10.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
setuptools==75.1.0
six==1.16.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
Werkzeug==3.0.4
wheel==0.44.0
